+++ src/resolvers/index.js
@@ -1,5 +1,6 @@
 const { GraphQLScalarType } = require('graphql');
 const Hra = require('../models/hra');
+const User = require('../models/user');
 const Mutation = require('./mutation');
 const Query = require('./query');
 const Subscription = require('./subscription');
@@ -35,6 +36,29 @@ const resolvers = {
         return hra;
       }
       return hra;
+    },
+    company: async parent => {
+      const user = await User.findById(parent.company);
+
+      if (!user) {
+        // eslint-disable-next-line no-shadow
+        const user = null;
+        return user;
+      }
+      return user;
+    },
+  },
+  UserWithCount: {
+    __resolveType(obj, context, info){
+      if (obj.type) {
+        return 'User';
+      }
+
+      if (obj.totalCount) {
+        return 'AddCountToUser';
+      }
+
+      return null;
     },
   },
 };